fix: apply audit review fixes, add configurable fee rate, close index race
Correções de auditoria (crypto/segurança):
- ChangePlanner: envio parcial agora gera troco (antes esvaziava a
  wallet em fee); addressToScriptPubKey valida rede/witness version/
  tamanho do programa do endereço de destino
- PokemonPassphrase: sorteio com SecureRandom independente em vez de
  derivar da entropia do mnemonic (a "25ª palavra" era reconstruível
  por quem visse as palavras)
- ReceiveAddressService: aplica o tweak Taproot (BIP341) no endereço
  de recebimento BIP86 — antes mostrava a chave interna não-tweakada,
  diferente do que o scanner/envio realmente monitoram/gastam
- TaprootSighashCalculator: reescrito pra bater com o BIP341 (SHA256
  simples em vez de double, hash_type de 1 byte, campos ausentes em
  vez de zerados sem annex/script-path); validado contra o vetor de
  teste oficial (novo TaprootSighashCalculatorTest)
- Psbt: ByteBuffer.wrap sem LITTLE_ENDIAN fazia UnsignedTransaction.
  parse/TxIn.parse/TxOut.parse ler campos multi-byte na ordem errada
  (achado ao validar o teste acima)
- WalletStorage: load()/save() sem nenhuma sincronização — dois
  load-modifica-save concorrentes (ex.: reservar índice de troco vs.
  refresh de saldo em background) podiam se sobrescrever. Adicionado
  lock + reserveNextInternalIndex()/reserveNextExternalIndex() pra
  tornar "ler índice atual + incrementar + persistir" atômico

Taxa de mineração escolhível no envio:
- Slider (0.5 sat/vB até 2x a taxa de prioridade alta da mempool,
  sugestão = taxa alta) com estimativa de tempo de confirmação
  (FeeTimeEstimator; abaixo de 1 sat/vB sempre avisa "mais de 1
  semana"); FeeEstimator/ChangePlanner aceitam taxa fracionária

UI: botão "Base" da navegação renomeado pra "Home"; saldo atualiza
ao tocar em Home (além do ciclo automático de 60s já existente).

+++ src/main/java/com/pokewallet/commands/SendCommand.kt
@@ -37,7 +37,7 @@ object SendCommand {
 
         // ── Fee rate ──────────────────────────────────────────────
         val fees     = BlockstreamClient.getFeeEstimates(network)
-        val feeRate  = fees.halfHour.toLong().coerceAtLeast(1L)
+        val feeRate  = fees.halfHour.coerceAtLeast(1.0)
 
         // ── Build spendable UTXO list ─────────────────────────────
         data class SpendableUtxo(
